Added Rcpp as a dependency
This is a hack: the package the FUSION uses to read in PLINK files,
namely R2plink, is not a proper package with a well-formed DESCRIPTION
file. On Ubuntu at least, plink2R::read_plink will terminate with an
error message about Rcpp availability if this dependency is not added.

Medium term solution: move away from plink2R.