Extended gene prediction function to return matched SNPs
Revised the original function for predicting genetically determined
expression to also return a logical vector to indicate which SNPs in
the reference data carry EQTL information (used for identifying loci
downstream). Not pretty, but both results need iteration over the
on-disk files with EQTL information, so probably better for all.